Fix the issues

+++ Udacity_frontend_store_project/mystore-frontend/src/app/components/cart/cart.ts
@@ -3,10 +3,11 @@ import { CartService } from '../../services/cartService';
 import { Router } from '@angular/router';
 import { CommonModule } from '@angular/common';
 import { RouterModule } from '@angular/router';
+import { CartItemComponent } from '../cart-item/cart-item';
 
 @Component({
   selector: 'app-cart',
-  imports: [CommonModule, RouterModule],
+  imports: [CommonModule, RouterModule, CartItemComponent],
   templateUrl: './cart.html',
   styleUrl: './cart.css',
 })
@@ -25,10 +26,14 @@ export class CartComponent implements OnInit {
   }
 
   removeFromCart(productId: number): void {
+  const confirmed = confirm('Are you sure you want to remove this item from the cart?');
+  if (confirmed) {
     this.cartService.removeFromCart(productId);
     this.cart = this.cartService.getCart();
     this.total = this.cartService.getTotal();
+    alert('Item removed from cart.');
   }
+}
 
   checkout(): void {
     this.router.navigate(['/checkout']);
